+++ src/account/TokenNftAllowance.js
@@ -15,6 +15,7 @@ export default class TokenNftAllowance {
      * @param {object} props
      * @param {TokenId} props.tokenId
      * @param {AccountId} props.spenderAccountId
+     * @param {AccountId | null} props.ownerAccountId
      * @param {Long[] | null} props.serialNumbers
      */
     constructor(props) {
@@ -31,6 +32,13 @@ export default class TokenNftAllowance {
          * @readonly
          */
         this.spenderAccountId = props.spenderAccountId;
+
+        /**
+         * The account ID of the owner of the hbar allowance.
+         *
+         * @readonly
+         */
+        this.ownerAccountId = props.ownerAccountId;
 
         /**
          * The current balance of the spender's token allowance.
@@ -57,6 +65,12 @@ export default class TokenNftAllowance {
             spenderAccountId: AccountId._fromProtobuf(
                 /** @type {proto.IAccountID} */ (allowance.spender)
             ),
+            ownerAccountId:
+                allowance.owner != null
+                    ? AccountId._fromProtobuf(
+                          /**@type {proto.IAccountID}*/ (allowance.owner)
+                      )
+                    : null,
             serialNumbers:
                 allowance.approvedForAll != null &&
                 allowance.approvedForAll.value
@@ -77,6 +91,10 @@ export default class TokenNftAllowance {
         return {
             tokenId: this.tokenId._toProtobuf(),
             spender: this.spenderAccountId._toProtobuf(),
+            owner:
+                this.ownerAccountId != null
+                    ? this.ownerAccountId._toProtobuf()
+                    : null,
             approvedForAll: this.serialNumbers == null ? { value: true } : null,
             serialNumbers: this.serialNumbers,
         };