+++ test/unit/ContractFunctionParameters.js
@@ -1,10 +1,10 @@
+import { ContractFunctionParameters } from "../../src/exports.js";
+
 import BigNumber from "bignumber.js";
-import * as ContractFunctionParameters from "../../src/contract/ContractFunctionParameters.js";
 
 describe("ContractFunctionParameters", function () {
     it("should convert number to BigNumber in addUint256()", function () {
-        const contractFunctionParameters =
-            new ContractFunctionParameters.default();
+        const contractFunctionParameters = new ContractFunctionParameters();
 
         const num = 111;
         const cfp = contractFunctionParameters.addUint256(num);
@@ -16,8 +16,7 @@ describe("ContractFunctionParameters", function () {
     });
 
     it("should convert number to BigNumber in addUint256Array()", function () {
-        const contractFunctionParameters =
-            new ContractFunctionParameters.default();
+        const contractFunctionParameters = new ContractFunctionParameters();
 
         const numArray = [111, 1112];